+++ src/control_plane/workload/immutable_project_application_options.rs
@@ -8,6 +8,7 @@ pub(crate) struct ImmutableProjectApplicationOptions<'operation> {
     pub(crate) installation_id: &'operation str,
     pub(crate) schema_version: u32,
     pub(crate) platform: &'operation str,
+    pub(crate) container_user: &'operation str,
     pub(crate) network_name: &'operation str,
     pub(crate) internal_http_port: u16,
 }