+++ src/control_plane/workload/project_process_operation_options.rs
@@ -11,5 +11,6 @@ pub(crate) struct ProjectProcessOperationOptions<'operation> {
     pub(crate) installation_id: &'operation str,
     pub(crate) schema_version: u32,
     pub(crate) platform: &'operation str,
+    pub(crate) container_user: &'operation str,
     pub(crate) network_name: &'operation str,
 }