+++ src/control_plane/workload/project_process_request_options.rs
@@ -6,4 +6,5 @@ pub(crate) struct ProjectProcessRequestOptions {
     pub(crate) plan: ProjectProcessPlan,
     pub(crate) metadata: ManagedResourceMetadata,
     pub(crate) platform: String,
+    pub(crate) container_user: String,
 }